+++ xenon-common/src/main/java/com/vmware/xenon/services/common/NodeGroupUtils.java
@@ -100,7 +100,7 @@ public class NodeGroupUtils {
             availableNodeCount++;
         }
 
-        if (availableNodeCount < selfNode.synchQuorum) {
+        if (availableNodeCount < Math.max(selfNode.synchQuorum, selfNode.membershipQuorum)) {
             return false;
         }
         return true;