+++ CHANGELOG.md
@@ -9,6 +9,14 @@
   factory service instance which can then be started with
   ServiceHost.startService()
 
+* Enable update of synch quorum through PATCH to
+  node-groups/<group>/config. Membership quorum was already
+  supported, now synch quorum is included.
+
+* Fix consensus protocol commit behavior. Commit messages
+  are no longer sent if the request to owner service fails
+  or the service does not have ENFORCE_QUORUM.
+
 * Split Service.handleDelete into Service.handleStop and
   Service.handleDelete. If the service is being stopped, the
   host has always added a special pragma, indicating this is